requirements.txt
transformers
torch
torchvision
Pillow
jiwer
datasets
matplotlib
sentencepiece
jupyter
protobuf
opencv-python
numpy
pymupdf
pandas
huggingface_hub
gradio